refactor: Split large service files to comply with 300-line limit
- Extract setting definitions to src/config/setting-definitions.ts (191 lines)
- Refactor SettingDefinitionService to import definitions (240 -> 47 lines, 80% reduction)
- Split GoogleSheetsService (646 lines) into specialized services:
  - google-sheets-metadata.service.ts (186 lines): Spreadsheet/sheet metadata operations
  - google-sheets-schema.service.ts (251 lines): Sheet creation and validation
  - google-sheets-data.service.ts (263 lines): Data CRUD operations
  - google-sheets.service.ts (153 lines): Facade pattern for unified interface
- Extract column utility to src/utils/column.ts (20 lines)
- All files now comply with 300-line maximum rule
