+++ src/services/google-sheets.service.ts
@@ -1,10 +1,14 @@
 /**
- * Google Sheets Service
+ * Google Sheets Service (Facade)
  *
- * Handles Google Sheets API operations using fetch-based REST API
- * Compatible with Cloudflare Workers runtime
+ * Unified interface for Google Sheets operations
+ * Delegates to specialized services:
+ * - GoogleSheetsMetadataService: Spreadsheet and sheet metadata
+ * - GoogleSheetsSchemaService: Sheet creation and validation
+ * - GoogleSheetsDataService: Data CRUD operations
  *
  * Uses Google Drive API v3 and Google Sheets API v4
+ * Compatible with Cloudflare Workers runtime
  */
 
 import type {
@@ -12,30 +16,22 @@ import type {
   SpreadsheetInfo,
   ValidationResult,
 } from '../types/google';
-
-/**
- * Convert 1-based column index to Excel-style column letter(s)
- * Examples: 1 -> A, 26 -> Z, 27 -> AA, 52 -> AZ, 703 -> AAA
- */
-function columnIndexToLetter(columnIndex: number): string {
-  let letter = '';
-  let n = columnIndex;
-
-  while (n > 0) {
-    const remainder = (n - 1) % 26;
-    letter = String.fromCharCode(65 + remainder) + letter;
-    n = Math.floor((n - 1) / 26);
-  }
-
-  return letter;
-}
+import { GoogleSheetsMetadataService } from './google-sheets-metadata.service';
+import { GoogleSheetsSchemaService } from './google-sheets-schema.service';
+import { GoogleSheetsDataService } from './google-sheets-data.service';
 
 export class GoogleSheetsService {
-  private accessToken: string;
+  private metadataService: GoogleSheetsMetadataService;
+  private schemaService: GoogleSheetsSchemaService;
+  private dataService: GoogleSheetsDataService;
 
   constructor(accessToken: string) {
-    this.accessToken = accessToken;
+    this.metadataService = new GoogleSheetsMetadataService(accessToken);
+    this.schemaService = new GoogleSheetsSchemaService(accessToken);
+    this.dataService = new GoogleSheetsDataService(accessToken);
   }
+
+  // === Metadata Operations ===
 
   /**
    * List available spreadsheets from Google Drive
@@ -43,72 +39,7 @@ export class GoogleSheetsService {
    * @returns Array of spreadsheet metadata
    */
   async listSpreadsheets(): Promise<SpreadsheetMetadata[]> {
-    const allFiles: Array<{
-      id: string;
-      name: string;
-      webViewLink: string;
-      createdTime: string;
-      modifiedTime: string;
-    }> = [];
-
-    let pageToken: string | undefined;
-
-    // Implement pagination to retrieve all files
-    do {
-      const params = new URLSearchParams({
-        q: "mimeType='application/vnd.google-apps.spreadsheet'",
-        fields: 'files(id,name,webViewLink,createdTime,modifiedTime),nextPageToken',
-        orderBy: 'modifiedTime desc',
-        pageSize: '100',
-      });
-
-      if (pageToken) {
-        params.set('pageToken', pageToken);
-      }
-
-      const response = await fetch(
-        `https://www.googleapis.com/drive/v3/files?${params}`,
-        {
-          headers: {
-            Authorization: `Bearer ${this.accessToken}`,
-            Accept: 'application/json',
-          },
-        }
-      );
-
-      if (!response.ok) {
-        const error = await response.text();
-        console.error('[GoogleSheetsService] API Error:', response.status, error);
-        throw new Error(
-          `Failed to list spreadsheets: ${response.status} ${error}`
-        );
-      }
-
-      const data = (await response.json()) as {
-        files?: Array<{
-          id: string;
-          name: string;
-          webViewLink: string;
-          createdTime: string;
-          modifiedTime: string;
-        }>;
-        nextPageToken?: string;
-      };
-
-      if (data.files) {
-        allFiles.push(...data.files);
-      }
-
-      pageToken = data.nextPageToken;
-    } while (pageToken);
-
-    return allFiles.map((file) => ({
-      id: file.id,
-      name: file.name,
-      url: file.webViewLink,
-      createdTime: file.createdTime,
-      modifiedTime: file.modifiedTime,
-    }));
+    return this.metadataService.listSpreadsheets();
   }
 
   /**
@@ -118,55 +49,7 @@ export class GoogleSheetsService {
    * @returns Spreadsheet information with sheets
    */
   async getSpreadsheetMetadata(sheetId: string): Promise<SpreadsheetInfo> {
-    const params = new URLSearchParams({
-      fields:
-        'spreadsheetId,properties.title,sheets(properties(sheetId,title,index,gridProperties))',
-    });
-
-    const response = await fetch(
-      `https://sheets.googleapis.com/v4/spreadsheets/${sheetId}?${params}`,
-      {
-        headers: {
-          Authorization: `Bearer ${this.accessToken}`,
-          Accept: 'application/json',
-        },
-      }
-    );
-
-    if (!response.ok) {
-      const error = await response.text();
-      throw new Error(
-        `Failed to get spreadsheet metadata: ${response.status} ${error}`
-      );
-    }
-
-    const data = (await response.json()) as {
-      spreadsheetId: string;
-      properties: { title: string };
-      sheets: Array<{
-        properties: {
-          sheetId: number;
-          title: string;
-          index: number;
-          gridProperties: {
-            rowCount: number;
-            columnCount: number;
-          };
-        };
-      }>;
-    };
-
-    return {
-      id: data.spreadsheetId,
-      name: data.properties.title,
-      sheets: data.sheets.map((sheet) => ({
-        sheetId: sheet.properties.sheetId,
-        title: sheet.properties.title,
-        index: sheet.properties.index,
-        rowCount: sheet.properties.gridProperties.rowCount,
-        columnCount: sheet.properties.gridProperties.columnCount,
-      })),
-    };
+    return this.metadataService.getSpreadsheetMetadata(sheetId);
   }
 
   /**
@@ -177,28 +60,10 @@ export class GoogleSheetsService {
    * @returns True if sheet exists
    */
   async sheetExists(sheetId: string, sheetTitle: string): Promise<boolean> {
-    const metadata = await this.getSpreadsheetMetadata(sheetId);
-    return metadata.sheets.some((s) => s.title === sheetTitle);
+    return this.metadataService.sheetExists(sheetId, sheetTitle);
   }
 
-  /**
-   * Get sheet ID by title
-   *
-   * @param spreadsheetId - Spreadsheet ID
-   * @param sheetTitle - Sheet title
-   * @returns Sheet ID (numeric)
-   */
-  private async getSheetIdByTitle(
-    spreadsheetId: string,
-    sheetTitle: string
-  ): Promise<number> {
-    const metadata = await this.getSpreadsheetMetadata(spreadsheetId);
-    const sheet = metadata.sheets.find((s) => s.title === sheetTitle);
-    if (!sheet) {
-      throw new Error(`Sheet "${sheetTitle}" not found`);
-    }
-    return sheet.sheetId;
-  }
+  // === Schema Operations ===
 
   /**
    * Create a new sheet with headers and column definitions
@@ -214,130 +79,12 @@ export class GoogleSheetsService {
     headers: string[],
     columnDefs: Record<string, unknown>[]
   ): Promise<void> {
-    console.log(`[GoogleSheetsService] Creating sheet "${sheetTitle}" with ${headers.length} columns...`);
-
-    // 1. Create new sheet with frozen rows for header (row 1) and column defs (row 2)
-    console.log(`[GoogleSheetsService] Step 1/3: Creating sheet structure...`);
-    const createRequest = {
-      requests: [
-        {
-          addSheet: {
-            properties: {
-              title: sheetTitle,
-              gridProperties: {
-                rowCount: 1000,
-                columnCount: headers.length,
-                frozenRowCount: 2, // Freeze both header and column definition rows
-              },
-            },
-          },
-        },
-      ],
-    };
-
-    const createResponse = await fetch(
-      `https://sheets.googleapis.com/v4/spreadsheets/${spreadsheetId}:batchUpdate`,
-      {
-        method: 'POST',
-        headers: {
-          Authorization: `Bearer ${this.accessToken}`,
-          'Content-Type': 'application/json',
-        },
-        body: JSON.stringify(createRequest),
-      }
+    return this.schemaService.createSheetWithHeaders(
+      spreadsheetId,
+      sheetTitle,
+      headers,
+      columnDefs
     );
-
-    if (!createResponse.ok) {
-      const error = await createResponse.text();
-      throw new Error(`Failed to create sheet: ${createResponse.status} ${error}`);
-    }
-    console.log(`[GoogleSheetsService] ✓ Sheet structure created`);
-
-    // 2. Set header values (row 1) and column definitions (row 2)
-    console.log(`[GoogleSheetsService] Step 2/3: Writing headers and column definitions...`);
-    const columnLetter = columnIndexToLetter(headers.length);
-    const columnDefStrings = columnDefs.map((def) => JSON.stringify(def));
-
-    const updateResponse = await fetch(
-      `https://sheets.googleapis.com/v4/spreadsheets/${spreadsheetId}/values/${sheetTitle}!A1:${columnLetter}2?valueInputOption=RAW`,
-      {
-        method: 'PUT',
-        headers: {
-          Authorization: `Bearer ${this.accessToken}`,
-          'Content-Type': 'application/json',
-        },
-        body: JSON.stringify({
-          values: [headers, columnDefStrings], // Row 1: headers, Row 2: column defs
-        }),
-      }
-    );
-
-    if (!updateResponse.ok) {
-      const error = await updateResponse.text();
-      throw new Error(`Failed to set headers and column definitions: ${updateResponse.status} ${error}`);
-    }
-    console.log(`[GoogleSheetsService] ✓ Headers and column definitions written`);
-
-    // 3. Format header row (row 1) and column definition row (row 2)
-    console.log(`[GoogleSheetsService] Step 3/3: Applying formatting...`);
-    const sheetId = await this.getSheetIdByTitle(spreadsheetId, sheetTitle);
-    const formatRequest = {
-      requests: [
-        // Format row 1 (headers) - bold with gray background
-        {
-          repeatCell: {
-            range: {
-              sheetId: sheetId,
-              startRowIndex: 0,
-              endRowIndex: 1,
-            },
-            cell: {
-              userEnteredFormat: {
-                backgroundColor: { red: 0.9, green: 0.9, blue: 0.9 },
-                textFormat: { bold: true },
-              },
-            },
-            fields: 'userEnteredFormat(backgroundColor,textFormat)',
-          },
-        },
-        // Format row 2 (column definitions) - light blue background
-        {
-          repeatCell: {
-            range: {
-              sheetId: sheetId,
-              startRowIndex: 1,
-              endRowIndex: 2,
-            },
-            cell: {
-              userEnteredFormat: {
-                backgroundColor: { red: 0.85, green: 0.92, blue: 0.95 },
-                textFormat: { fontSize: 9 },
-              },
-            },
-            fields: 'userEnteredFormat(backgroundColor,textFormat)',
-          },
-        },
-      ],
-    };
-
-    const formatResponse = await fetch(
-      `https://sheets.googleapis.com/v4/spreadsheets/${spreadsheetId}:batchUpdate`,
-      {
-        method: 'POST',
-        headers: {
-          Authorization: `Bearer ${this.accessToken}`,
-          'Content-Type': 'application/json',
-        },
-        body: JSON.stringify(formatRequest),
-      }
-    );
-
-    if (!formatResponse.ok) {
-      const error = await formatResponse.text();
-      throw new Error(`Failed to format headers: ${formatResponse.status} ${error}`);
-    }
-    console.log(`[GoogleSheetsService] ✓ Formatting applied successfully`);
-    console.log(`[GoogleSheetsService] ✓✓ Sheet "${sheetTitle}" created successfully`);
   }
 
   /**
@@ -347,69 +94,10 @@ export class GoogleSheetsService {
    * @returns Validation result
    */
   async validateSheetStructure(sheetId: string): Promise<ValidationResult> {
-    const metadata = await this.getSpreadsheetMetadata(sheetId);
-
-    const usersSheet = metadata.sheets.find((s) => s.title === '_Users');
-    const rolesSheet = metadata.sheets.find((s) => s.title === '_Roles');
-    const filesSheet = metadata.sheets.find((s) => s.title === '_Files');
-
-    const errors: string[] = [];
-    const warnings: string[] = [];
-
-    if (!usersSheet) {
-      errors.push('Required sheet "_Users" not found');
-    }
-    if (!rolesSheet) {
-      errors.push('Required sheet "_Roles" not found');
-    }
-    if (!filesSheet) {
-      errors.push('Required sheet "_Files" not found');
-    }
-
-    // Validate _Users columns if sheet exists
-    if (usersSheet) {
-      const response = await fetch(
-        `https://sheets.googleapis.com/v4/spreadsheets/${sheetId}/values/_Users!A1:Z1`,
-        {
-          headers: {
-            Authorization: `Bearer ${this.accessToken}`,
-            Accept: 'application/json',
-          },
-        }
-      );
-
-      if (response.ok) {
-        const data = (await response.json()) as {
-          values?: string[][];
-        };
-        const columns = data.values?.[0] || [];
-
-        const requiredColumns = [
-          'object_id',
-          'username',
-          '_password_hash',
-          'email',
-          'name',
-          'status',
-          'created_at',
-        ];
-        for (const col of requiredColumns) {
-          if (!columns.includes(col)) {
-            errors.push(`_Users sheet missing required column: ${col}`);
-          }
-        }
-      }
-    }
-
-    return {
-      valid: errors.length === 0,
-      errors,
-      warnings,
-      hasUsersSheet: !!usersSheet,
-      hasRolesSheet: !!rolesSheet,
-      hasFilesSheet: !!filesSheet,
-    };
+    return this.schemaService.validateSheetStructure(sheetId);
   }
+
+  // === Data Operations ===
 
   /**
    * Get all data from a sheet as an array of objects with column names as keys
@@ -423,62 +111,7 @@ export class GoogleSheetsService {
     sheetTitle: string,
     options?: { includePrivateColumns?: boolean }
   ): Promise<Array<Record<string, string | number | boolean>>> {
-    // Get headers (row 1) and all data (from row 3 onwards)
-    const headersResponse = await fetch(
-      `https://sheets.googleapis.com/v4/spreadsheets/${spreadsheetId}/values/${sheetTitle}!1:1`,
-      {
-        headers: {
-          Authorization: `Bearer ${this.accessToken}`,
-          Accept: 'application/json',
-        },
-      }
-    );
-
-    if (!headersResponse.ok) {
-      const error = await headersResponse.text();
-      throw new Error(`Failed to get sheet headers: ${headersResponse.status} ${error}`);
-    }
-
-    const headersData = (await headersResponse.json()) as {
-      values?: Array<Array<string>>;
-    };
-
-    const headers = headersData.values?.[0] || [];
-
-    // Get all data rows (starting from row 3 to skip header and column definition rows)
-    const dataResponse = await fetch(
-      `https://sheets.googleapis.com/v4/spreadsheets/${spreadsheetId}/values/${sheetTitle}!A3:${columnIndexToLetter(headers.length)}`,
-      {
-        headers: {
-          Authorization: `Bearer ${this.accessToken}`,
-          Accept: 'application/json',
-        },
-      }
-    );
-
-    if (!dataResponse.ok) {
-      const error = await dataResponse.text();
-      throw new Error(`Failed to get sheet data: ${dataResponse.status} ${error}`);
-    }
-
-    const data = (await dataResponse.json()) as {
-      values?: Array<Array<string | number | boolean>>;
-    };
-
-    const rows = data.values || [];
-
-    // Convert rows to objects using headers as keys
-    // Exclude columns starting with _ unless includePrivateColumns is true
-    const includePrivate = options?.includePrivateColumns ?? false;
-    return rows.map((row) => {
-      const rowObject: Record<string, string | number | boolean> = {};
-      headers.forEach((header, index) => {
-        if (includePrivate || !header.startsWith('_')) {
-          rowObject[header] = row[index] ?? '';
-        }
-      });
-      return rowObject;
-    });
+    return this.dataService.getSheetData(spreadsheetId, sheetTitle, options);
   }
 
   /**
@@ -495,90 +128,12 @@ export class GoogleSheetsService {
     objectId: string,
     rowData: Record<string, string | number | boolean>
   ): Promise<void> {
-    // Get headers to know column order
-    const headersResponse = await fetch(
-      `https://sheets.googleapis.com/v4/spreadsheets/${spreadsheetId}/values/${sheetTitle}!1:1`,
-      {
-        headers: {
-          Authorization: `Bearer ${this.accessToken}`,
-          Accept: 'application/json',
-        },
-      }
+    return this.dataService.updateRow(
+      spreadsheetId,
+      sheetTitle,
+      objectId,
+      rowData
     );
-
-    if (!headersResponse.ok) {
-      const error = await headersResponse.text();
-      throw new Error(`Failed to get sheet headers: ${headersResponse.status} ${error}`);
-    }
-
-    const headersData = (await headersResponse.json()) as {
-      values?: Array<Array<string>>;
-    };
-
-    const headers = headersData.values?.[0] || [];
-
-    // Get all data to find the row with matching object_id
-    const allDataResponse = await fetch(
-      `https://sheets.googleapis.com/v4/spreadsheets/${spreadsheetId}/values/${sheetTitle}!A3:${columnIndexToLetter(headers.length)}`,
-      {
-        headers: {
-          Authorization: `Bearer ${this.accessToken}`,
-          Accept: 'application/json',
-        },
-      }
-    );
-
-    if (!allDataResponse.ok) {
-      const error = await allDataResponse.text();
-      throw new Error(`Failed to get sheet data: ${allDataResponse.status} ${error}`);
-    }
-
-    const allData = (await allDataResponse.json()) as {
-      values?: Array<Array<string | number | boolean>>;
-    };
-
-    const rows = allData.values || [];
-
-    // Find row index by object_id (first column)
-    const rowIndex = rows.findIndex((row) => row[0] === objectId);
-
-    if (rowIndex === -1) {
-      throw new Error(`Row with object_id "${objectId}" not found in sheet "${sheetTitle}"`);
-    }
-
-    // Calculate actual row number (add 3: 1 for header, 1 for column defs, 1 for 1-based indexing)
-    const actualRowNumber = rowIndex + 3;
-
-    // Build values array in the correct column order
-    const values = headers.map((header) => {
-      if (header in rowData) {
-        return rowData[header];
-      }
-      // Keep existing value if not provided in rowData
-      const existingRow = rows[rowIndex];
-      const columnIndex = headers.indexOf(header);
-      return existingRow[columnIndex] ?? '';
-    });
-
-    const columnLetter = columnIndexToLetter(values.length);
-    const response = await fetch(
-      `https://sheets.googleapis.com/v4/spreadsheets/${spreadsheetId}/values/${sheetTitle}!A${actualRowNumber}:${columnLetter}${actualRowNumber}?valueInputOption=RAW`,
-      {
-        method: 'PUT',
-        headers: {
-          Authorization: `Bearer ${this.accessToken}`,
-          'Content-Type': 'application/json',
-        },
-        body: JSON.stringify({
-          values: [values],
-        }),
-      }
-    );
-
-    if (!response.ok) {
-      const error = await response.text();
-      throw new Error(`Failed to update row: ${response.status} ${error}`);
-    }
   }
 
   /**
@@ -593,54 +148,6 @@ export class GoogleSheetsService {
     sheetTitle: string,
     rowData: Record<string, string | number | boolean>
   ): Promise<void> {
-    // Get headers to know column order
-    const headersResponse = await fetch(
-      `https://sheets.googleapis.com/v4/spreadsheets/${spreadsheetId}/values/${sheetTitle}!1:1`,
-      {
-        headers: {
-          Authorization: `Bearer ${this.accessToken}`,
-          Accept: 'application/json',
-        },
-      }
-    );
-
-    if (!headersResponse.ok) {
-      const error = await headersResponse.text();
-      throw new Error(`Failed to get sheet headers: ${headersResponse.status} ${error}`);
-    }
-
-    const headersData = (await headersResponse.json()) as {
-      values?: Array<Array<string>>;
-    };
-
-    const headers = headersData.values?.[0] || [];
-
-    // Build values array in the correct column order
-    const values = headers.map((header) => {
-      if (header in rowData) {
-        return rowData[header];
-      }
-      // Return empty string for missing columns
-      return '';
-    });
-
-    const response = await fetch(
-      `https://sheets.googleapis.com/v4/spreadsheets/${spreadsheetId}/values/${sheetTitle}!A:${columnIndexToLetter(headers.length)}:append?valueInputOption=RAW`,
-      {
-        method: 'POST',
-        headers: {
-          Authorization: `Bearer ${this.accessToken}`,
-          'Content-Type': 'application/json',
-        },
-        body: JSON.stringify({
-          values: [values],
-        }),
-      }
-    );
-
-    if (!response.ok) {
-      const error = await response.text();
-      throw new Error(`Failed to append row: ${response.status} ${error}`);
-    }
+    return this.dataService.appendRow(spreadsheetId, sheetTitle, rowData);
   }
 }
